requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
requests==2.31.0
Pillow==10.2.0
psycopg2-binary==2.9.9   # si usas PostgreSQL (puedes quitarlo si usas SQLite)
